habanalabs: register to pci shutdown callback
Git-commit: c78cff56baad24ebb10d748e1a1b78bae203debe
References: bsn#19
Patch-mainline: v5.10.9

[ Upstream commit fcaebc7354188b0d708c79df4390fbabd4d9799d ]

We need to make sure our device is idle when rebooting a virtual
machine. This is done in the driver level.

The firmware will later handle FLR but we want to be extra safe and
stop the devices until the FLR is handled.

Signed-off-by: Oded Gabbay <ogabbay@kernel.org>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.9-0071-habanalabs-register-to-pci-shutdown-callback.patch